Properly rename and drop morphToMany relationship fieldtypes

+++ app/Observers/FieldObserver.php
@@ -82,13 +82,33 @@ class FieldObserver
             $table = $container->getTable();
 
             if ($old['handle'] !== $new['handle']) {
-                $fieldtype = fieldtypes()->get($new['type']);
-                $column    = $fieldtype->getColumn('type');
+                $fieldtype    = fieldtypes()->get($new['type']);
+                $column       = $fieldtype->getColumn('type');
+                $relationship = $fieldtype->getRelationship();
 
                 if (! is_null($column)) {
                     Schema::table($table, function ($table) use ($old, $new) {
                         $table->renameColumn("`{$old['handle']}`", "`{$new['handle']}`");
                     });
+                }
+
+                if (! is_null($relationship)) {
+                    switch($relationship) {
+                        case 'morphToMany':
+                            $oldTableName = Str::plural(Str::addAbleSuffix($old['handle']));
+                            $newTableName = Str::plural(Str::addAbleSuffix($new['handle']));
+
+                            if (! Schema::hasTable($newTableName)) {
+                                Schema::rename($oldTableName, $newTableName);
+                                Schema::table($newTableName, function($table) use ($old, $new) {
+                                    $table->renameColumn(Str::singular($old['handle']).'_id', Str::singular($new['handle']).'_id');
+                                    $table->renameColumn(Str::addAbleSuffix($old['handle']).'_id', Str::addAbleSuffix($new['handle']).'_id');
+                                    $table->renameColumn(Str::addAbleSuffix($old['handle']).'_type', Str::addAbleSuffix($new['handle']).'_type');
+                                });
+                            }
+
+                            break;
+                    }
                 }
             }
 
@@ -118,16 +138,36 @@ class FieldObserver
     {
         $fieldset   = $field->section->fieldset;
         $containers = $this->getFieldsettables($fieldset);
+        
+        $fieldtype    = fieldtypes()->get($field->type);
+        $relationship = $fieldtype->getRelationship();
+        $column       = $fieldtype->getColumn('type');
+        $settings     = $fieldtype->getColumn('settings') ?? [];
 
-        $containers->each(function($container) use ($field) {
-            $table = $container->getTable();
+        array_unshift($settings, $field->handle);
 
-            if (Schema::hasColumn($table, $field->handle)) {
-                Schema::table($table, function ($table) use ($field) {
-                    $table->dropColumn($field->handle);
-                });
+        if (! is_null($column)) {
+            $containers->each(function($container) use ($field) {
+                $table = $container->getTable();
+
+                if (Schema::hasColumn($table, $field->handle)) {
+                    Schema::table($table, function ($table) use ($field) {
+                        $table->dropColumn($field->handle);
+                    });
+                }
+            });
+        }
+
+        if (! is_null($relationship)) {
+            switch($relationship) {
+                case 'morphToMany':
+                    $tableName = Str::plural(Str::addAbleSuffix($field->handle));
+
+                    Schema::dropIfExists($tableName);
+
+                    break;
             }
-        });
+        }
     }
 
     /**
